Refactor tests for encrypted storage.
